Make the location in TcLclEnv and CtLoc into a RealSrcSpan
Previously it was a SrcSpan, which can be an UnhelpulSrcSpan,
but actually for TcLclEnv and CtLoc we always know it is
a real source location, and it's good to make the types
reflect that fact.

There is a continuing slight awkwardness (not new with this
patch) about what "file name" to use for GHCi code.  Current
we say "<interactive>" which seems just about OK.

Cherry-Picked-From: d2b6e7671e823fce0c0fbb2bed739fa948a23219
Cherry-Picked-By: Niklas Hambüchen <niklas@fpcomplete.com>

Changes for the cherry-pick:

* RealSrcLoc for external core (removed in 7.10) was set to
  "<external core>".