Use IP based CallStack in error and undefined
Summary:
This is a proof-of-concept implementation of using https://phabricator.haskell.org/D578 (Eric Seidel's "Expose source locations via Implicit Parameters of type GHC.Location.Location" patch) in the `error` and `undefined` functions in base.

The feature he implemented allows us to obtain the current source code location (in the same format as the one you get when your program dies with a pattern match error) via an ImplicitParameter that you can pull out of thin air (because it is a GHC-inbuilt one).

Further, it does not only provide the current location, but also builds up a "CallStack" containing the caller function if the ImplicitParameter is present in the type signature and the caller of *that* function if that function also has it in the type signature, all the way up to the first function that doesn't have it in the signature.

This makes it possible to see where a function was called, without requiring any kind of support from the caller.

The feature has two main motivations:

1) Easy way to obtain code location without need of TemplateHaskell (e.g. for logging libraries)
2) Providing source location information when programs crash due to partial functions (`error`, `undefined` etc.)

This patch implements point 2.

(You can find Eric's intent to do this at https://mail.haskell.org/pipermail/ghc-devs/2015-March/008645.html.)

This patch was developed as part of my work for FP Complete; we would like to use it in a client project that requires good error messages in case of a crash -- and hopefully everywhere else as well! I also have a strong personal interest in making one of our biggest dreads (partial functions) less scary (with better error messages than "Exception: Prelude.undefined").

I call this patch a proof-of-concept because:

* It's modifying base.
* Due to strong restrictions on what can import what in base, I had to put the CallStack-enabled `error` into Prelude instead of `GHC.Base`; I'd like to see if there's a way around that.
* This changes the type of the `error` function, which may be part of the language standard (one might argue though that it it does not practically change the type since the implicit CallStack parameter is provided "magically" by GHC if the caller doesn't do it, so for callers that don't know about the CallStack implicit parameter, the type change is invisible).

I hope to resolve these points by discssion on this patch.

Reviewers: austin, hvr

Subscribers: thomie

Differential Revision: https://phabricator.haskell.org/D861