Fix hWaitForInput not being interruptible (#8684).
Summary:
* Make fdReady a `ccall interruptible`
* Move EINTR loop from C into Haskell so that Haskell code
  can run between interruptions
* Fix `ccall interruptible` not being interruptible when calling
  `fdReady()` on Windows, by using an Event HANDLE to signal
  interruption
* Call `interruptible` between `fdReady()` calls so that
  exceptions have a chance to be delivered.
* Call `yield` on the non-threaded RTS between `fdReady()` calls
  so that functions like `timeout` get a chance to produce their
  exceptions.

This also fixes Ctrl+C not terminating cancelling Haskell
programs that are stuck in `hWaitForInput`.
Until now, one had to force a "dirty exit" with double-Ctrl+C;
with this commit, this is no longer necessary.

I've verified this improvement on Linux and Windows.

I've also added commentary in `handle_tick()` that explains the
reasoning behind this, as an outcome of the discussion on
https://phabricator.haskell.org/D42.

Reviewers: hvr, bgamari, erikd, simonmar

Subscribers: rwbarton, thomie

Differential Revision: https://phabricator.haskell.org/D42