rts: Refactor: Move USE_PTHREAD_FOR_ITIMER definition to a header file.
This allows us to use that value elsewhere.

Specifically the intent is to use it to decide how to interrupt blocking
IO for context-switching in the non-threaded RTS; for that we have to
know by which method the context switch timer is implemented
(to decide whether `pthread_kill()` should be used or not).